+++ CHANGELOG.md
@@ -4,10 +4,9 @@ Mapbox welcomes participation and contributions from everyone.
 
 ## main
 * Improve stability of symbol placement when using `FollowPuckViewportState`.
-
 * Expose `clusterMinPoints` property for `GeoJSONSource` and for `ClusterOptions`
-
 * Root properties (`Atmosphere`, `Lights`, `Projection`, `Terrain`, `Transition`) are now revertible for all styles. 
+* Introduce raster particles rendering example
 
 ## 11.4.0 - 22 May, 2024